[gc] Fix 64-bit platform warning: src/gc/gc_ims.c:854: warning: cast to pointer from integer of different size
git-svn-id: https://svn.parrot.org/parrot/trunk@26796 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/gc/gc_ims.c
@@ -851,7 +851,7 @@ parrot_gc_ims_collect(PARROT_INTERP, int check_only)
         Parrot_dod_profile_start(interp);
     g_ims = (Gc_ims_private *)arena_base->gc_private;
     ret   = Parrot_forall_header_pools(interp, POOL_BUFFER,
-            (void*)check_only, collect_cb);
+            (void*)(long)check_only, collect_cb);
     if (ret)
         return ret;
     if (check_only)
